Phase A (Step 5 prep): widen sys_sbrk to u64 ABI on the SYSCALL fast path
To prepare for moving the program / DLIB region into the upper half,
brk addresses also need to escape the u32 ceiling. Add a parallel u64
ABI for SYS_SBRK so the kernel can return brk addresses anywhere in
the canonical-low half.

  kernel/src/syscall/handlers/process.rs
    - new pub fn sys_sbrk_u64(increment: i64) -> u64
      operates on full u64 brk values from the start; checks DLIB_REGION_START
      and HEAP_LIMIT directly as u64 instead of truncating to u32 first
    - existing sys_sbrk(i32) -> u32 now delegates to sys_sbrk_u64 and
      truncates at the boundary (kept as a fallback for any in-kernel
      caller that still goes through dispatch_inner)

  kernel/src/syscall/mod.rs
    - syscall_dispatch_64 special-cases SYS_SBRK and routes to
      sys_sbrk_u64

  libs/libsyscall/src/lib.rs
    - sbrk()'s failure check changed from `u32::MAX as u64` to `u64::MAX`
  libs/stdlib/src/process.rs
    - sbrk() now uses syscall1_u64() and treats `usize::MAX` as the
      failure marker so callers can keep the Rust-pointer-sized `usize`
      return type
  libs/stdlib/src/heap.rs
    - heap init failure check follows: `brk == usize::MAX as u64`

No address has actually moved yet — brk still lives in the low half
(programs at 0x0800_0000, DLIBs at 0x0400_0000+). The u64 ABI is
purely future-proofing for the upcoming user_vmap layout shift into
the upper canonical-low half.

Verified: build green, image boots in QEMU/KVM through compositor
spawn, no exceptions.

+++ libs/stdlib/src/heap.rs
@@ -101,7 +101,7 @@ pub fn init() {
     }
 
     let brk = crate::process::sbrk(0) as u64;
-    if brk == u32::MAX as u64 {
+    if brk == usize::MAX as u64 {
         return;
     }
     unsafe {
